+++ flame/config.yaml
@@ -2,7 +2,7 @@ admin_email: manuel.pastor@upf.edu
 admin_name: Manuel Pastor
 config_status: true
 homepage: http://phi.upf.edu
-model_repository_path: C:\Users\mpastor\Documents\soft\flame\flame\models
-space_repository_path: C:\Users\mpastor\Documents\soft\flame\flame\spaces
-predictions_repository_path: C:\Users\mpastor\Documents\soft\flame\flame\predictions
+model_repository_path: D:\synced\internals\bsep
+space_repository_path: D:\synced\internals\bsep
+predictions_repository_path: D:\synced\internals\bsep
 provider: UPF (Pompeu Fabra University)
